requirements.txt
openai
Flask
google-cloud-bigquery
python-dotenv
pytest
pytest-flask
pytest-mock